requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
google-genai
scikit-learn
numpy
pydantic